Completed reconstruction of Login
