+++ frontend/src/util/inputUtil.ts
@@ -1,6 +1,20 @@
 /**
  * This file contains the utilities needed by forms
  */
+export interface IFields {
+  label: string, 
+  id: string,
+  type?: string,
+  value: any,
+}
+
+export interface IButtons {
+  text: string, 
+  type?: string, 
+  to?: string,  // note: this will only redirect to home
+  bootstrapClass: string,
+}
+
 
 export const updateField = (e: React.BaseSyntheticEvent, index: number, form: any) => {
   let elementValue = (e.target as HTMLInputElement).value;
@@ -9,6 +23,7 @@ export const updateField = (e: React.BaseSyntheticEvent, index: number, form: an
   return formCopy;
 }
 
-export const submitVerify = () => {
+export function submitVerify(form: IFields[]) {
+  return form.every((item: any) => item.value.length > 0)
+}
 
-}
